Create stateless component
